final touches

+++ controllers/contact.controller.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contact.model");
+
 //@description Get All Contacts
 //@route GET /api/contacts
 //@access private
@@ -17,7 +18,12 @@ const createContact = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("All fields are required!");
   }
-  const newContact = await Contact.create({ name, email, phone });
+  const newContact = await Contact.create({
+    name,
+    email,
+    phone,
+    user_id: req.user.id,
+  });
   res.status(201).json(newContact);
 });
 
@@ -42,6 +48,12 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("contact not found");
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "user doesn't have permission to edit another user's contact"
+    );
+  }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -60,6 +72,12 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("contact not found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "user doesn't have permission to edit another user's contact"
+    );
   }
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
